ccitt: add the AutoDetectHeight decoding option
Before this commit, the decodeEOL method would usually not advance the
bitReader, if the expected EOL code was not found. But in some
circumstances, it would still advance if it saw a valid decoding mode,
since EOL-decoding and mode-decoding was combined into a single table,
even though EOL was not listed in Table 1 of the "ITU-T Recommendation
T.6" spec. After this commit, EOL's and modes are separate.

Change-Id: If4b83baf8083abdf14462576c56aaeb7cc8ff0f9
Reviewed-on: https://go-review.googlesource.com/c/image/+/211238
Reviewed-by: Benny Siegert <bsiegert@gmail.com>